# spec/openapi/ai_agent_platform_openapi_v1.yaml
openapi: 3.1.0
info:
  title: MSSQL Analysis Agent Platform API
  version: 0.1.0
  summary: OpenAPI contract skeleton for the MSSQL analysis, documentation, and Java/MyBatis draft-generation platform.
  description: |
    Starter API contract for the central Agent platform that analyzes MSSQL Stored Procedures
    and related database metadata, produces documentation, and creates draft Java/MyBatis
    conversion artifacts.

    This contract is intentionally skeleton-level. It defines the stable API surface and
    response shapes expected by the first implementation wave while preserving project
    policies:

    - metadata access is read-only and routed through the MSSQL Metadata MCP boundary;
    - generated documentation, Java/MyBatis code, DTO/VO/Model, and DDL are draft artifacts;
    - default P25 workflows stop at validation-complete draft output without user review UI;
    - deferred publish-like actions still require validation evidence and retained approval capability;
    - live row-data queries, automatic DDL execution, and unreviewed code application are out of scope.
servers:
  - url: http://localhost:{port}
    description: Local API/BFF server
    variables:
      port:
        default: "8000"
        description: APP_PORT resolved by Makefile or scripts/resolve_dev_ports.sh.
...
paths:
...
  /api/v1/requests/sp-analysis:
    post:
      tags: [requests]
      summary: Submit an MSSQL Stored Procedure analysis request.
      operationId: createStoredProcedureAnalysisRequest
      description: |
        Starts a draft-only analysis workflow for one stored procedure. The resulting job is
        expected to collect metadata through the MCP boundary, build a canonical analysis
        model, generate requested artifacts, and stop at the validation-complete draft gate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SPAnalysisRequest"
            examples:
              spWrapper:
                value:
                  dbProfileId: plf
                  target:
                    type: PROCEDURE
                    schema: dbo
                    name: usp_OrderRequest_Select
                  outputs:
                    - SP_ANALYSIS_DOCUMENT
                    - DEPENDENCY_REPORT
                    - JAVA_MYBATIS_DRAFT
                  options:
                    includeModernizationHints: true
                    includeEvidenceRefs: true
                    useAiToolOrchestration: true
      responses:
        "202":
          description: Request accepted and workflow job created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitRequestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/jobs/{jobId}:
    get:
      tags: [jobs]
      summary: Read workflow job status.
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Current job state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/jobs/{jobId}/artifacts:
    get:
      tags: [artifacts]
      summary: List draft artifacts created by a job.
      operationId: listJobArtifacts
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Draft artifacts for the job.
          content:
            application/json:
              schema:
                type: object
                required: [jobId, artifacts]
                properties:
                  jobId:
                    type: string
                  artifacts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ArtifactSummary"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/artifacts/{artifactId}/validation:
    post:
      tags: [artifacts]
      summary: Validate a draft artifact before review or approval.
      operationId: validateArtifact
      parameters:
        - $ref: "#/components/parameters/ArtifactId"
      responses:
        "200":
          description: Validation report for the artifact.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationReport"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/metadata/search:
    get:
      tags: [metadata]
      summary: Search read-only MSSQL metadata objects.
      operationId: searchMetadataObjects
      description: |
        Searches metadata object identities through the MSSQL Metadata MCP read-only boundary.
        The response is limited to object identity, profile/database context, snapshot/evidence
        references, caveats, review-required markers, and blocker codes. It never returns row
        data, executes stored procedures, exposes SQL definition text, or performs DDL/DML.
        PPM search must not fall back to PLF when PPM is unavailable.
      parameters:
        - name: dbProfileId
          in: query
          required: true
          schema:
            type: string
          examples:
            master:
              value: master
            ppm:
              value: ppm
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          example: order
        - name: objectTypes
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: "#/components/schemas/MetadataSearchObjectType"
            default: [PROCEDURE, TABLE, VIEW, FUNCTION]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Metadata object search results and blocker/caveat context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataSearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "424":
          $ref: "#/components/responses/DependencyBlocked"
        "503":
          $ref: "#/components/responses/DependencyBlocked"
...
components:
  parameters:
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
      example: job_123
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema:
        type: string
      example: art_123
  responses:
    BadRequest:
      description: Invalid request payload or unsupported workflow option.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The requested read-only metadata dependency is blocked by policy or access.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    DependencyBlocked:
      description: Required metadata dependency is unavailable or blocked by policy.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    TargetObject:
      type: object
      required: [type, schema, name]
      properties:
        type:
          type: string
          enum: [PROCEDURE, TABLE, VIEW, FUNCTION]
        schema:
          type: string
          examples: [dbo]
        name:
          type: string
          examples: [usp_OrderRequest_Select]
    SPAnalysisRequest:
      type: object
      required: [dbProfileId, target, outputs]
      properties:
        dbProfileId:
          type: string
          description: Metadata profile identifier, for example plf or master.
        target:
          $ref: "#/components/schemas/TargetObject"
        outputs:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RequestedOutputType"
        options:
          $ref: "#/components/schemas/SPAnalysisOptions"
    SPAnalysisOptions:
      type: object
      properties:
        includeEvidenceRefs:
          type: boolean
          default: true
        includeModernizationHints:
          type: boolean
          default: true
        useLlmAnalysis:
          type: boolean
          default: true
        llmProfileId:
          type: string
          default: openai_sp_semantic_analysis
          enum:
            - openai_sp_semantic_analysis
            - openai_fast_test
        allowSpDefinitionToModel:
          type: boolean
          default: true
        useAiToolOrchestration:
          type: boolean
          default: true
    SubmitRequestResponse:
      type: object
      required: [requestId, jobId, status]
      properties:
        requestId:
          type: string
        jobId:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        echo:
          type: object
          description: Optional starter implementation echo of the normalized request.
    Job:
      type: object
      required: [jobId, requestId, status]
      properties:
        jobId:
          type: string
        requestId:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        currentStep:
          oneOf:
            - $ref: "#/components/schemas/WorkflowStepType"
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        progress:
          type: number
          minimum: 0
          maximum: 1
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchBlocker"
        caveats:
          type: array
          items:
            type: string
        failureReason:
          type: string
    JobStatus:
      type: string
      description: VALIDATION_COMPLETE is the default P25 terminal draft status after validation; review and approval statuses remain for deferred compatibility.
      enum:
        - SUBMITTED
        - COLLECTING_METADATA
        - ANALYZING
        - GENERATING
        - VALIDATING
        - VALIDATION_COMPLETE
        - REVIEW_PENDING
        - APPROVED
        - REJECTED
        - PUBLISHED
        - FAILED
        - CANCELED
    WorkflowStepType:
      type: string
      enum:
        - COLLECT_METADATA
        - ANALYZE
        - GENERATE
        - VALIDATE
        - REVIEW
        - PUBLISH
    ArtifactSummary:
      type: object
      required: [artifactId, type, status]
      properties:
        artifactId:
          type: string
        jobId:
          type: string
        type:
          $ref: "#/components/schemas/ArtifactType"
        status:
          $ref: "#/components/schemas/ArtifactStatus"
        title:
          type: string
        evidenceCoverage:
          type: number
          minimum: 0
          maximum: 1
...
    ArtifactType:
      type: string
      description: Persisted draft artifact type. Request output groups map to one or more of these values.
      enum:
        - SP_ANALYSIS_DOC
        - DEPENDENCY_REPORT
        - METADATA_QUERY_RESULT
        - SCHEMA_ENRICHMENT_RESULT
        - MAPPER_XML
        - MAPPER_INTERFACE
        - SERVICE_DRAFT
        - DTO_DRAFT
        - VO_DRAFT
        - MODEL_DRAFT
        - DDL_DRAFT
        - VALIDATION_REPORT
        - APPROVAL_LOG
    RequestedOutputType:
      type: string
      description: User-facing output group requested at submission time.
      enum:
        - SP_ANALYSIS_DOCUMENT
        - DEPENDENCY_REPORT
        - TABLE_COLUMN_METADATA
        - JAVA_MYBATIS_DRAFT
        - DTO_MODEL_DRAFT
        - DDL_DRAFT
    ArtifactStatus:
      type: string
      enum:
        - DRAFT
        - VALIDATED
        - REVIEW_PENDING
        - APPROVED
        - REJECTED
        - PUBLISHED
        - ARCHIVED
    EvidenceRef:
      type: object
      required: [type, objectRef, locator]
      properties:
        type:
          type: string
          enum: [MSSQL_METADATA, STATIC_ANALYSIS, LLM_INFERENCE, POLICY, TEMPLATE, USER_INPUT]
        objectRef:
          type: string
          examples: [plf.dbo.usp_OrderRequest_Select]
        locator:
          type: string
        snapshotId:
          type: string
    ValidationReport:
      type: object
      required: [artifactId, status, checks]
      properties:
        validationReportId:
          type: string
        artifactId:
          type: string
        status:
          type: string
          enum: [PASSED, FAILED, REVIEW_REQUIRED]
        checks:
          type: array
          items:
            $ref: "#/components/schemas/ValidationCheck"
        missingEvidence:
          type: array
          items:
            type: string
        manualReviewPoints:
          type: array
          items:
            type: string
    ValidationCheck:
      type: object
      required: [ruleId, severity, result]
      properties:
        ruleId:
          type: string
        severity:
          type: string
          enum: [INFO, WARNING, ERROR, BLOCKER]
        result:
          type: string
          enum: [PASS, FAIL, REVIEW_REQUIRED]
        message:
          type: string
...
    MetadataSearchObjectType:
      type: string
      enum: [PROCEDURE, TABLE, VIEW, FUNCTION]
    MetadataObjectIdentity:
      type: object
      required: [schema, name, type]
      properties:
        schema:
          type: string
        name:
          type: string
        type:
          $ref: "#/components/schemas/MetadataSearchObjectType"
    MetadataSearchBlocker:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples:
            - PPM_MANIFEST_TEMPLATE_ONLY
            - METADATA_SEARCH_MCP_TOOL_MISSING
            - PPM_DB_NOT_FOUND
            - PPM_DB_ACCESS_DENIED
            - LIVE_METADATA_UNAVAILABLE
            - DEPENDENCY_METADATA_INCOMPLETE
            - P21_LIVE_PPM_REQUIRED
        message:
          type: string
    MetadataSearchResult:
      type: object
      required:
        - objectIdentity
        - sourceProfile
        - sourceDatabase
        - evidenceRefs
        - caveats
        - reviewRequired
        - blockers
      properties:
        objectIdentity:
          $ref: "#/components/schemas/MetadataObjectIdentity"
        sourceProfile:
          type: string
        sourceDatabase:
          type: string
        snapshotId:
          type: string
        evidenceRefs:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceRef"
        caveats:
          type: array
          items:
            type: string
        reviewRequired:
          type: boolean
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchBlocker"
    MetadataSearchResponse:
      type: object
      required:
        - dbProfileId
        - query
        - objectTypes
        - limit
        - sourceProfile
        - sourceDatabase
        - results
        - caveats
        - reviewRequired
        - blockers
      properties:
        dbProfileId:
          type: string
        query:
          type: string
        objectTypes:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchObjectType"
        limit:
          type: integer
          minimum: 1
          maximum: 100
        sourceProfile:
          type: string
        sourceDatabase:
          type: string
        snapshotId:
          type: string
        collectedAt:
          type: string
          format: date-time
        results:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchResult"
        caveats:
          type: array
          items:
            type: string
        reviewRequired:
          type: boolean
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchBlocker"
...
    ErrorResponse:
      type: object
      required: [detail, code]
      properties:
        detail:
          type: string
        code:
          type: string
          examples:
            - DEPENDENCY_BLOCKED
            - P21_LIVE_PORTAL_REQUIRED_ENV_MISSING
            - P21_LIVE_PLF_UNAVAILABLE
            - P21_LIVE_PPM_REQUIRED
            - PPM_DB_ACCESS_DENIED
            - PPM_DB_NOT_FOUND